+++ Projects/Server/Network/Packets/Outgoing/AttributeNormalizer.cs
