+++ Projects/Server/Network/Packets/Outgoing/VendorBuyPackets.cs
