+++ Projects/UOContent/Items/Misc/BulletinBoards.cs
@@ -229,7 +229,7 @@ namespace Server.Items
 
         public static void Initialize()
         {
-            PacketHandlers.Register(0x71, 0, true, BBClientRequest);
+            Packets.Register(0x71, 0, true, BBClientRequest);
         }
 
         public static void BBClientRequest(NetState state, CircularBufferReader reader)
